fix stop-music functionality

+++ play_music.sh
@@ -20,6 +20,10 @@ then
 elif [ "$1" = "stop-music" ];
 then
 	music_pid="$(ps -ef | grep mpv | grep '\--loop --no-video' | awk '{print $2}')"
+	if [ -z "$music_pid" ]; then
+		echo 'no music is being played'
+		exit 0
+	fi
 	kill $music_pid
 	exit 0
 fi
